% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppfiles-package.R
\docType{package}
\name{cppfiles-package}
\alias{cppfiles}
\alias{cppfiles-package}
\title{cppfiles: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
